no longer export indexing stuff and rename file
