+++ src/RcppExports.cpp
@@ -23051,105 +23051,6 @@ BEGIN_RCPP
     return R_NilValue;
 END_RCPP
 }
-// cpp_torch_tensor_index
-Rcpp::XPtr<XPtrTorchTensor> cpp_torch_tensor_index(Rcpp::XPtr<XPtrTorchTensor> self, Rcpp::XPtr<XPtrTorchTensorIndex> index);
-RcppExport SEXP _torch_cpp_torch_tensor_index(SEXP selfSEXP, SEXP indexSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< Rcpp::XPtr<XPtrTorchTensor> >::type self(selfSEXP);
-    Rcpp::traits::input_parameter< Rcpp::XPtr<XPtrTorchTensorIndex> >::type index(indexSEXP);
-    rcpp_result_gen = Rcpp::wrap(cpp_torch_tensor_index(self, index));
-    return rcpp_result_gen;
-END_RCPP
-}
-// cpp_torch_tensor_index_new
-Rcpp::XPtr<XPtrTorchTensorIndex> cpp_torch_tensor_index_new();
-RcppExport SEXP _torch_cpp_torch_tensor_index_new() {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    rcpp_result_gen = Rcpp::wrap(cpp_torch_tensor_index_new());
-    return rcpp_result_gen;
-END_RCPP
-}
-// cpp_torch_tensor_index_append_tensor
-void cpp_torch_tensor_index_append_tensor(Rcpp::XPtr<XPtrTorchTensorIndex> self, Rcpp::XPtr<XPtrTorchTensor> x);
-RcppExport SEXP _torch_cpp_torch_tensor_index_append_tensor(SEXP selfSEXP, SEXP xSEXP) {
-BEGIN_RCPP
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< Rcpp::XPtr<XPtrTorchTensorIndex> >::type self(selfSEXP);
-    Rcpp::traits::input_parameter< Rcpp::XPtr<XPtrTorchTensor> >::type x(xSEXP);
-    cpp_torch_tensor_index_append_tensor(self, x);
-    return R_NilValue;
-END_RCPP
-}
-// cpp_torch_tensor_index_append_bool
-void cpp_torch_tensor_index_append_bool(Rcpp::XPtr<XPtrTorchTensorIndex> self, bool x);
-RcppExport SEXP _torch_cpp_torch_tensor_index_append_bool(SEXP selfSEXP, SEXP xSEXP) {
-BEGIN_RCPP
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< Rcpp::XPtr<XPtrTorchTensorIndex> >::type self(selfSEXP);
-    Rcpp::traits::input_parameter< bool >::type x(xSEXP);
-    cpp_torch_tensor_index_append_bool(self, x);
-    return R_NilValue;
-END_RCPP
-}
-// cpp_torch_tensor_index_append_int64
-void cpp_torch_tensor_index_append_int64(Rcpp::XPtr<XPtrTorchTensorIndex> self, int64_t x);
-RcppExport SEXP _torch_cpp_torch_tensor_index_append_int64(SEXP selfSEXP, SEXP xSEXP) {
-BEGIN_RCPP
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< Rcpp::XPtr<XPtrTorchTensorIndex> >::type self(selfSEXP);
-    Rcpp::traits::input_parameter< int64_t >::type x(xSEXP);
-    cpp_torch_tensor_index_append_int64(self, x);
-    return R_NilValue;
-END_RCPP
-}
-// cpp_torch_tensor_index_append_ellipsis
-void cpp_torch_tensor_index_append_ellipsis(Rcpp::XPtr<XPtrTorchTensorIndex> self);
-RcppExport SEXP _torch_cpp_torch_tensor_index_append_ellipsis(SEXP selfSEXP) {
-BEGIN_RCPP
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< Rcpp::XPtr<XPtrTorchTensorIndex> >::type self(selfSEXP);
-    cpp_torch_tensor_index_append_ellipsis(self);
-    return R_NilValue;
-END_RCPP
-}
-// cpp_torch_tensor_index_append_none
-void cpp_torch_tensor_index_append_none(Rcpp::XPtr<XPtrTorchTensorIndex> self);
-RcppExport SEXP _torch_cpp_torch_tensor_index_append_none(SEXP selfSEXP) {
-BEGIN_RCPP
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< Rcpp::XPtr<XPtrTorchTensorIndex> >::type self(selfSEXP);
-    cpp_torch_tensor_index_append_none(self);
-    return R_NilValue;
-END_RCPP
-}
-// cpp_torch_tensor_index_append_slice
-void cpp_torch_tensor_index_append_slice(Rcpp::XPtr<XPtrTorchTensorIndex> self, Rcpp::XPtr<XPtrTorch> x);
-RcppExport SEXP _torch_cpp_torch_tensor_index_append_slice(SEXP selfSEXP, SEXP xSEXP) {
-BEGIN_RCPP
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< Rcpp::XPtr<XPtrTorchTensorIndex> >::type self(selfSEXP);
-    Rcpp::traits::input_parameter< Rcpp::XPtr<XPtrTorch> >::type x(xSEXP);
-    cpp_torch_tensor_index_append_slice(self, x);
-    return R_NilValue;
-END_RCPP
-}
-// cpp_torch_slice
-Rcpp::XPtr<XPtrTorchSlice> cpp_torch_slice(Rcpp::XPtr<XPtrTorchoptional_int64_t> start, Rcpp::XPtr<XPtrTorchoptional_int64_t> end, Rcpp::XPtr<XPtrTorchoptional_int64_t> step);
-RcppExport SEXP _torch_cpp_torch_slice(SEXP startSEXP, SEXP endSEXP, SEXP stepSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< Rcpp::XPtr<XPtrTorchoptional_int64_t> >::type start(startSEXP);
-    Rcpp::traits::input_parameter< Rcpp::XPtr<XPtrTorchoptional_int64_t> >::type end(endSEXP);
-    Rcpp::traits::input_parameter< Rcpp::XPtr<XPtrTorchoptional_int64_t> >::type step(stepSEXP);
-    rcpp_result_gen = Rcpp::wrap(cpp_torch_slice(start, end, step));
-    return rcpp_result_gen;
-END_RCPP
-}
 // cpp_lantern_configure
 void cpp_lantern_configure(bool log);
 RcppExport SEXP _torch_cpp_lantern_configure(SEXP logSEXP) {
@@ -25591,15 +25492,6 @@ static const R_CallMethodDef CallEntries[] = {
     {"_torch_evaluate_slices", (DL_FUNC) &_torch_evaluate_slices, 2},
     {"_torch_Tensor_slice", (DL_FUNC) &_torch_Tensor_slice, 4},
     {"_torch_Tensor_slice_put", (DL_FUNC) &_torch_Tensor_slice_put, 4},
-    {"_torch_cpp_torch_tensor_index", (DL_FUNC) &_torch_cpp_torch_tensor_index, 2},
-    {"_torch_cpp_torch_tensor_index_new", (DL_FUNC) &_torch_cpp_torch_tensor_index_new, 0},
-    {"_torch_cpp_torch_tensor_index_append_tensor", (DL_FUNC) &_torch_cpp_torch_tensor_index_append_tensor, 2},
-    {"_torch_cpp_torch_tensor_index_append_bool", (DL_FUNC) &_torch_cpp_torch_tensor_index_append_bool, 2},
-    {"_torch_cpp_torch_tensor_index_append_int64", (DL_FUNC) &_torch_cpp_torch_tensor_index_append_int64, 2},
-    {"_torch_cpp_torch_tensor_index_append_ellipsis", (DL_FUNC) &_torch_cpp_torch_tensor_index_append_ellipsis, 1},
-    {"_torch_cpp_torch_tensor_index_append_none", (DL_FUNC) &_torch_cpp_torch_tensor_index_append_none, 1},
-    {"_torch_cpp_torch_tensor_index_append_slice", (DL_FUNC) &_torch_cpp_torch_tensor_index_append_slice, 2},
-    {"_torch_cpp_torch_slice", (DL_FUNC) &_torch_cpp_torch_slice, 3},
     {"_torch_cpp_lantern_configure", (DL_FUNC) &_torch_cpp_lantern_configure, 1},
     {"_torch_cpp_lantern_version", (DL_FUNC) &_torch_cpp_lantern_version, 0},
     {"_torch_cpp_lantern_init", (DL_FUNC) &_torch_cpp_lantern_init, 1},
